fixed some use of int which broke when files exceed 2GB, including getting content-length header as a string and parsing manually as a long changed method sig in ProgressListener to take long instead of int
git-svn-id: https://opencadc.googlecode.com/svn/trunk@1122 728ff76a-78ac-11de-a72b-d90af8dea425

+++ projects/cadcUtil/src/ca/nrc/cadc/net/HttpTransfer.java
@@ -286,7 +286,7 @@ public abstract class HttpTransfer implements Runnable
      * @throws IOException
      * @throws InterruptedException
      */
-    protected void ioLoop(InputStream istream, OutputStream ostream, int sz, int startingPos)
+    protected void ioLoop(InputStream istream, OutputStream ostream, int sz, long startingPos)
         throws IOException, InterruptedException
     {
         log.debug("ioLoop: using java.io with byte[] buffer size " + sz);
@@ -294,7 +294,7 @@ public abstract class HttpTransfer implements Runnable
 
         int nb = 0;
         int nb2 = 0;
-        int tot = startingPos; // non-zero for resumed transfer
+        long tot = startingPos; // non-zero for resumed transfer
         int n = 0;
 
         if (progressListener != null)
@@ -337,7 +337,7 @@ public abstract class HttpTransfer implements Runnable
      * @throws IOException
      * @throws InterruptedException
      */
-    protected void nioLoop(InputStream istream, OutputStream ostream, int sz, int startingPos)
+    protected void nioLoop(InputStream istream, OutputStream ostream, int sz, long startingPos)
         throws IOException, InterruptedException
     {
         log.debug("[Download] nioLoop: using java.nio with ByteBuffer size " + sz);
@@ -348,7 +348,7 @@ public abstract class HttpTransfer implements Runnable
         ReadableByteChannel rbc = Channels.newChannel(istream);
         WritableByteChannel wbc = Channels.newChannel(ostream);
 
-        int tot = startingPos; // non-zero for resumed transfer
+        long tot = startingPos; // non-zero for resumed transfer
         int count = 0;
 
         ByteBuffer buffer = ByteBuffer.allocate(sz);
